Feature/parte1

+++ Parte 1 y 3/src/main/java/edu/eci/arst/concprg/prodcons/StartProduction.java
@@ -23,7 +23,7 @@ public class StartProduction {
         Queue<Integer> queue=new LinkedBlockingQueue<>();
         
         
-        new Producer(queue,Long.MAX_VALUE).start();
+        new Producer(queue,Integer.MAX_VALUE).start();
         
         //let the producer create products for 5 seconds (stock).
         try {
